Chore : 불필요한 유틸함수 삭제 및 통합 #74
