--- src/pages/list/utils/background/backgroundMap.js
@@ -1,7 +1,0 @@
-import image01 from '@/components/cardlist/assets/image_01.png';
-import image02 from '@/components/cardlist/assets/image_02.png';
-
-export const backgroundImageMap = {
-  'image_01.png': image01,
-  'image_02.png': image02,
-};
