--- src/pages/list/utils/background/colorMap.js
@@ -1,18 +1,0 @@
-export const backgroundColorMap = {
-  purple: {
-    color: 'var(--purple-200)',
-    pattern: '@/components/cardlist/assets/pattern_01.png',
-  },
-  beige: {
-    color: 'var(--beige-200)',
-    pattern: '@/components/cardlist/assets/pattern_02.png',
-  },
-  blue: {
-    color: 'var(--blue-200)',
-    pattern: '@/components/cardlist/assets/pattern_03.png',
-  },
-  green: {
-    color: 'var(--green-200)',
-    pattern: '@/components/cardlist/assets/pattern_04.png',
-  },
-};
