--- src/pages/list/utils/background/parseImageFilename.js
@@ -1,4 +1,0 @@
-export function parseImageFilename(raw) {
-  if (!raw) return null;
-  return raw.trim().split('/').pop();
-}
