--- src/pages/list/utils/background/selectBackgroundForCard.js
@@ -1,13 +1,0 @@
-import { parseImageFilename } from './parseImageFilename.js';
-import { resolveImageBackground } from './resolveImageBackground.js';
-import { resolveColorBackground } from './resolveColorBackground.js';
-
-export function selectBackgroundForCard(item) {
-  if (item.backgroundImageURL) {
-    const filename = parseImageFilename(item.backgroundImageURL);
-    const imageBackground = resolveImageBackground(filename);
-    if (imageBackground) return imageBackground;
-  }
-
-  return resolveColorBackground(item.backgroundColor);
-}
